feat(ubx): Move GPS NMEA support to a dedicated file
