+++ radio/src/gps.h
@@ -53,4 +53,6 @@ void gpsWakeup();
 // Send a 0-terminated frame
 void gpsSendFrame(const char * frame);
 
+uint32_t GPS_coord_to_degrees(const char * coordinateString);
+
 #endif // _GPS_H_
